chore(deps): update sn_messaging, sn_data_types BREAKING CHANGE: removes send_message_to_client api,
