+++ Cargo.toml
@@ -25,12 +25,8 @@ rand_chacha = "~0.2.2"
 thiserror = "1.0.23"
 xor_name = "1.1.0"
 resource_proof = "0.8.0"
-# sn_messaging = "~4.0.0"
-sn_messaging = { git = "https://github.com/oetyng/sn_messaging", branch = "all-acc-in-routing" }
-# sn_messaging = { path = "../sn_messaging" }
-# sn_data_types = "~0.14.3"
-sn_data_types = { git = "https://github.com/oetyng/sn_data_types", branch = "remove-msgenvelope" }
-# sn_data_types = { path = "../sn_data_types" }
+sn_messaging = "~5.0.0"
+sn_data_types = "~0.15.0"
 
   [dependencies.bls]
   package = "threshold_crypto"
@@ -73,3 +69,4 @@ yansi = "~0.5.0"
   [dev-dependencies.tokio]
   version = "~0.2.24"
   features = [ "stream", "udp" ]
+
