+++ src/routing/approved.rs
@@ -6,7 +6,10 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::{Command, SplitBarrier};
+use super::{
+    enduser_registry::{EndUserRegistry, SocketId},
+    Command, SplitBarrier,
+};
 use crate::{
     consensus::{
         DkgCommands, DkgFailureProof, DkgFailureProofSet, DkgKey, DkgVoter, Proof, ProofShare,
@@ -38,7 +41,6 @@ use bytes::Bytes;
 use ed25519_dalek::Verifier;
 use itertools::Itertools;
 use resource_proof::ResourceProof;
-use sn_data_types::{PublicKey as EndUserPK, Signature as EndUserSig};
 use sn_messaging::{
     node::NodeMessage,
     section_info::{
@@ -46,74 +48,13 @@ use sn_messaging::{
     },
     DstLocation, EndUser, MessageType, SrcLocation,
 };
-use std::{
-    cmp,
-    collections::{btree_map::Entry, BTreeMap},
-    net::SocketAddr,
-    slice,
-};
+use std::{cmp, net::SocketAddr, slice};
 use tokio::sync::mpsc;
 use xor_name::{Prefix, XorName};
 
 pub(crate) const RESOURCE_PROOF_DATA_SIZE: usize = 64;
 pub(crate) const RESOURCE_PROOF_DIFFICULTY: u8 = 2;
 const KEY_CACHE_SIZE: u8 = 5;
-
-type SocketId = XorName;
-
-struct EndUserRegistry {
-    clients: BTreeMap<SocketAddr, EndUser>,
-    socket_id_mapping: BTreeMap<SocketId, SocketAddr>,
-}
-
-impl EndUserRegistry {
-    pub fn new() -> Self {
-        Self {
-            clients: BTreeMap::default(),
-            socket_id_mapping: BTreeMap::default(),
-        }
-    }
-
-    pub fn get_enduser_by_addr(&self, socketaddr: &SocketAddr) -> Option<EndUser> {
-        self.clients.get(socketaddr).copied()
-    }
-
-    pub fn get_socket_addr(&self, socket_id: &SocketId) -> Option<&SocketAddr> {
-        self.socket_id_mapping.get(socket_id)
-    }
-
-    pub fn try_add(
-        &mut self,
-        sender: SocketAddr,
-        end_user_pk: EndUserPK,
-        socketaddr_sig: EndUserSig,
-    ) -> Result<()> {
-        if let Ok(data) = &bincode::serialize(&sender) {
-            end_user_pk
-                .verify(&socketaddr_sig, data)
-                .map_err(|_e| Error::InvalidState)?;
-        } else {
-            return Err(Error::InvalidState);
-        }
-        let socket_id = if let Ok(socket_id_src) = &bincode::serialize(&socketaddr_sig) {
-            XorName::from_content(&[socket_id_src])
-        } else {
-            return Err(Error::InvalidState);
-        };
-        let end_user = EndUser::Client {
-            public_key: end_user_pk,
-            socket_id,
-        };
-        match self.socket_id_mapping.entry(socket_id) {
-            Entry::Vacant(_) => {
-                let _ = self.clients.insert(sender, end_user);
-                let _ = self.socket_id_mapping.insert(socket_id, sender);
-            }
-            Entry::Occupied(_) => (),
-        }
-        Ok(())
-    }
-}
 
 // The approved stage - node is a full member of a section and is performing its duties according
 // to its persona (adult or elder).
@@ -781,13 +722,7 @@ impl Approved {
     // If elder, always handle UserMessage, otherwise handle it only if addressed directly to us
     // as a node.
     fn should_handle_user_message(&self, dst: &DstLocation) -> bool {
-        let is_elder = self.is_elder();
-        let are_we_dst = if let DstLocation::Node(name) = dst {
-            name == &self.node.name()
-        } else {
-            false
-        };
-        is_elder || are_we_dst
+        self.is_elder() || dst == &DstLocation::Node(self.node.name())
     }
 
     // Decide how to handle a `Vote` message.
